Clean Garbage Collection after each Revolution
Memory maintenance is disabled between each Revolution. For obvious reasons of predictability of performance. 
If it is justified for the execution time. But it leads to an uncontrolled memory leak giving a false idea of it in reports, and a risk of exceeding the memory limit.

This correction forces a manual cleaning between each Revolution, in a way that does not impact the time measurement. The time measurement itself is placed a little more precisely in the code.